Closer to finished csvextract
No anonymous customer?
Also if a customer has multiple orders they will duplicate another line for a second order

+++ Source/CoffeePointOfSale/Services/Customer/Order.cs
@@ -4,9 +4,9 @@
     {
         public string CustomerId { get; set; }
         public string Date { get; set; }
-        public double Tax { get; set; }
-        public double Subtotal { get; set; }
-        public double Total { get; set; }
+        public decimal Tax { get; set; }
+        public decimal Subtotal { get; set; }
+        public decimal Total { get; set; }
         public string PaymentMethod { get; set; }
         public List<OrderedItem> OrderedItems { get; set; } = new();
     }
